+++ features/bootstrap/ApplicationContext.php
@@ -4,13 +4,14 @@ use Behat\Behat\Tester\Exception\PendingException;
 use Behat\Behat\Context\Context;
 use Behat\Gherkin\Node\PyStringNode;
 use Behat\Gherkin\Node\TableNode;
-use Fake\DialogHelper;
+use Fake\Prompter;
 use Fake\ReRunner;
 use Matcher\ApplicationOutputMatcher;
 use Matcher\ExitStatusMatcher;
 use Matcher\ValidJUnitXmlMatcher;
 use PhpSpec\Console\Application;
 use PhpSpec\Matcher\MatchersProviderInterface;
+use Symfony\Component\Console\Helper\HelperSet;
 use Symfony\Component\Console\Tester\ApplicationTester;
 
 /**
@@ -34,9 +35,9 @@ class ApplicationContext implements Context, MatchersProviderInterface
     private $tester;
 
     /**
-     * @var DialogHelper
+     * @var Prompter
      */
-    private $dialogHelper;
+    private $prompter;
 
     /**
      * @var ReRunner
@@ -53,16 +54,15 @@ class ApplicationContext implements Context, MatchersProviderInterface
 
         $this->tester = new ApplicationTester($this->application);
 
-        $this->setupDialogHelper();
         $this->setupReRunner();
+        $this->setupPrompter();
     }
 
-    private function setupDialogHelper()
+    private function setupPrompter()
     {
-        $this->dialogHelper = new DialogHelper();
+        $this->prompter = new Prompter();
 
-        $helperSet = $this->application->getHelperSet();
-        $helperSet->set($this->dialogHelper);
+        $this->application->getContainer()->set('console.prompter', $this->prompter);
     }
 
     private function setupReRunner()
@@ -120,7 +120,7 @@ class ApplicationContext implements Context, MatchersProviderInterface
 
         $this->addOptionToArguments($option, $arguments);
 
-        $this->dialogHelper->setAnswer($answer=='y');
+        $this->prompter->setAnswer($answer=='y');
 
         $this->lastExitCode = $this->tester->run($arguments, array('interactive' => true));
     }
@@ -154,7 +154,7 @@ class ApplicationContext implements Context, MatchersProviderInterface
      */
     public function iShouldBePromptedForCodeGeneration()
     {
-        expect($this->dialogHelper)->toHaveBeenAsked();
+        expect($this->prompter)->toHaveBeenAsked();
     }
 
     /**
@@ -162,7 +162,7 @@ class ApplicationContext implements Context, MatchersProviderInterface
      */
     public function iShouldNotBePromptedForCodeGeneration()
     {
-        expect($this->dialogHelper)->toNotHaveBeenAsked();
+        expect($this->prompter)->toNotHaveBeenAsked();
     }
 
     /**