+++ spec/PhpSpec/Console/IOSpec.php
@@ -2,6 +2,7 @@
 
 namespace spec\PhpSpec\Console;
 
+use PhpSpec\Console\Prompter;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 use PhpSpec\Config\OptionsConfig;
@@ -12,7 +13,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 class IOSpec extends ObjectBehavior
 {
-    function let(InputInterface $input, OutputInterface $output, DialogHelper $dialogHelper, OptionsConfig $config)
+    function let(InputInterface $input, OutputInterface $output, OptionsConfig $config, Prompter $prompter)
     {
         $input->isInteractive()->willReturn(true);
         $input->getOption('no-code-generation')->willReturn(false);
@@ -21,7 +22,7 @@ class IOSpec extends ObjectBehavior
         $config->isCodeGenerationEnabled()->willReturn(true);
         $config->isStopOnFailureEnabled()->willReturn(false);
 
-        $this->beConstructedWith($input, $output, $dialogHelper, $config);
+        $this->beConstructedWith($input, $output, $config, $prompter);
     }
 
     function it_has_io_interface()