+++ src/PhpSpec/Console/IO.php
@@ -15,9 +15,11 @@ namespace PhpSpec\Console;
 
 use PhpSpec\IO\IOInterface;
 use Symfony\Component\Console\Helper\DialogHelper;
+use Symfony\Component\Console\Helper\QuestionHelper;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use PhpSpec\Config\OptionsConfig;
+use Symfony\Component\Console\Question\ConfirmationQuestion;
 
 /**
  * Class IO deals with input and output from command line interaction
@@ -37,11 +39,6 @@ class IO implements IOInterface
      * @var \Symfony\Component\Console\Output\OutputInterface
      */
     private $output;
-
-    /**
-     * @var \Symfony\Component\Console\Helper\DialogHelper
-     */
-    private $dialogHelper;
 
     /**
      * @var string
@@ -64,17 +61,22 @@ class IO implements IOInterface
     private $consoleWidth;
 
     /**
+     * @var Prompter
+     */
+    private $prompter;
+
+    /**
      * @param InputInterface  $input
      * @param OutputInterface $output
-     * @param DialogHelper    $dialogHelper
      * @param OptionsConfig   $config
+     * @param Prompter        $prompter
      */
-    public function __construct(InputInterface $input, OutputInterface $output, DialogHelper $dialogHelper, OptionsConfig $config)
+    public function __construct(InputInterface $input, OutputInterface $output, OptionsConfig $config, Prompter $prompter)
     {
         $this->input   = $input;
         $this->output  = $output;
-        $this->dialogHelper = $dialogHelper;
         $this->config  = $config;
+        $this->prompter = $prompter;
     }
 
     /**
@@ -259,17 +261,6 @@ class IO implements IOInterface
     }
 
     /**
-     * @param string      $question
-     * @param string|null $default
-     *
-     * @return string
-     */
-    public function ask($question, $default = null)
-    {
-        return $this->dialogHelper->ask($this->output, $question, $default);
-    }
-
-    /**
      * @param string $question
      * @param bool   $default
      *
@@ -285,22 +276,9 @@ class IO implements IOInterface
         $lines[] = '<question>'.str_repeat(' ', $this->getBlockWidth() - 8).'</question> <value>'.
             ($default ? '[Y/n]' : '[y/N]').'</value> ';
 
-        return $this->dialogHelper->askConfirmation(
-            $this->output, implode("\n", $lines). "\n", $default
-        );
-    }
+        $formattedQuestion = implode("\n", $lines) . "\n";
 
-    /**
-     * @param string      $question
-     * @param callable    $validator
-     * @param int|false   $attempts
-     * @param string|null $default
-     *
-     * @return string
-     */
-    public function askAndValidate($question, $validator, $attempts = false, $default = null)
-    {
-        return $this->dialogHelper->askAndValidate($this->output, $question, $validator, $attempts, $default);
+        return $this->prompter->askConfirmation($formattedQuestion, $default);
     }
 
     /**